+++ src/Models/Country.php
@@ -72,4 +72,12 @@ class Country extends Model implements Localizable
     {
         $query->where('code', config('address.locality.country'));
     }
+
+    /**
+     * @return array<int, class-string>
+     */
+    protected static function localityDescendants(): array
+    {
+        return [config('address.models.state')];
+    }
 }